my commit file

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of char: %lu byte(s)\n", sizeof(a));
+	printf("Size of char: %lu byte(s)\n",(unsigned long) sizeof(a));
 	printf("Size of int: %lu byte(s)\n", sizeof(b));
 	printf("Size of long int: %lu byte(s)\n", sizeof(c));
 	printf("Size of long long int: %lu byte(s)\n", sizeof(d));
